Update js canvas examples

+++ examples/Canvas/Creatures.kt
@@ -147,10 +147,10 @@ class Creature(override var pos : Vector, val state : CanvasState) : Shape() {
         val p1 = tailPos + tailDirection.rotatedBy(angle) * tailSize
         val p2 = tailPos + tailDirection.rotatedBy(-angle) * tailSize
         val middlePoint = position + tailDirection * radius * 1.0
-        context.moveTo(tailPos.x, tailPos.y)
-        context.lineTo(p1.x, p1.y)
+        context.moveTo(tailPos.x.toInt(), tailPos.y.toInt())
+        context.lineTo(p1.x.toInt(), p1.y.toInt())
         context.quadraticCurveTo(middlePoint.x, middlePoint.y, p2.x, p2.y)
-        context.lineTo(tailPos.x, tailPos.y)
+        context.lineTo(tailPos.x.toInt(), tailPos.y.toInt())
     }
 
     fun drawEye(context : Context) {
@@ -246,10 +246,10 @@ class CanvasState(val canvas : Canvas) {
 
     fun clear() {
         context.fillStyle = "#FFFFFF"
-        context.fillRect(0.0, 0.0, width, height)
+        context.fillRect(0, 0, width.toInt(), height.toInt())
         context.strokeStyle = "#000000"
         context.lineWidth = 4.0
-        context.strokeRect(0.0, 0.0, width, height)
+        context.strokeRect(0, 0, width.toInt(), height.toInt())
     }
 
     fun draw() {
